requirements.txt
pillow==9.2.0
deflate==0.3.0